requirements.txt
pandas>=1.5.0
numpy>=1.20.0
yfinance>=0.2.0
pydantic>=1.10.0
PyYAML>=6.0
loguru>=0.6.0